Upgrade V8 to the 7.8 release branch
The build system of V8 has significantly changed. It now uses LLVM on
all platforms and a dedicated Debian (sid) based sysroot on Linux.  In
order to simplify the extension build process, it is now part of the V8
build system and uses GN. It comes with several enhancements like being
able to use C++14 in the extension. It also add support for Windows.