Manage teardown with shared_ptr not CountDownLatch
It's hard to manage object lifecycle in the C++ side of PyMiniRacer
because objects are often referenced in callbacks sent into the Isolate
message pump. Before we tried to explicitly manage lifecycle, and used
a CountDownLatch to ensure all the async work was done before closing
out the Isolate. In this change, we just switch everything to shared_ptr
and let the Isolate shut itself down when nothing else refers to it.